[SBI-644] [SBI-756] Исправить масштабирование и отображение фильтров, Исправить текст на карточках пользователей и добавить проверку на удаление (#445)
* make filters adaptive on team-card and streams pages

* update trackerlist text to Разблокировать/Удалить, add confirmation to deleting user, update styles

* fix tracker-page tests

* add onKeyDown adjacent to onClick div element

* remove unneccessery comment from label

* add stream-page tests for new conditions with showCheckboxes

* add tracker-page tests for new conditions with showCheckboxes

* add tracker-list-page tests for new conditions with window.confirm

+++ apps/frontend/src/accets/listuser/TrackerListPage.js
@@ -236,6 +236,9 @@ function TrackerListPage({ endpoint }) {
                       className="cancel-button"
                       onClick={(e) => {
                         e.stopPropagation();
+                        if (!window.confirm("Вы точно хотите удалить этого пользователя?")) {
+                          return;
+                        }
                         deleteUser(tracker.username);
                         setActiveMobileMenu(null);
                       }}
@@ -270,7 +273,7 @@ function TrackerListPage({ endpoint }) {
                       />
                     </button>
                     {hoveredButton === "cancel" && (
-                      <span className="tooltip tooltip-red">Отменить</span>
+                      <span className="tooltip tooltip-red">Удалить</span>
                     )}
 
                     {/* Отклонить */}
@@ -278,6 +281,9 @@ function TrackerListPage({ endpoint }) {
                       className="cancel-button"
                       onClick={(e) => {
                         e.stopPropagation();
+                        if (!window.confirm("Вы точно хотите удалить этого пользователя?")) {
+                          return;
+                        }
                         deleteUser(tracker.username);
                         setActiveMobileMenu(null);
                       }}
@@ -290,7 +296,7 @@ function TrackerListPage({ endpoint }) {
                       />
                     </button>
                     {hoveredButton === "confirm" && (
-                      <span className="tooltip tooltip-green">Подтвердить</span>
+                      <span className="tooltip tooltip-green">Разблокировать</span>
                     )}
                   </>
                 )}
